all camera implemented, added camera controls in the readme

+++ smolEngine/Cube.h
@@ -9,21 +9,21 @@ struct Cube : Graphic
 	static const int NUM_TRIANGLES = NUM_FACES * 2;
 	static const int NUM_VERTEXS = NUM_TRIANGLES * 3;
 
-	const vec3 a;
-	const vec3 b;
-	const vec3 c;
-	const vec3 d;
-	const vec3 e;
-	const vec3 f;
-	const vec3 g;
-	const vec3 h;
+	const dvec3 a;
+	const dvec3 b;
+	const dvec3 c;
+	const dvec3 d;
+	const dvec3 e;
+	const dvec3 f;
+	const dvec3 g;
+	const dvec3 h;
 
-	const vec3 red;
-	const vec3 green;
-	const vec3 blue;
-	const vec3 yellow;
-	const vec3 white;
-	const vec3 black;
+	const dvec3 red;
+	const dvec3 green;
+	const dvec3 blue;
+	const dvec3 yellow;
+	const dvec3 white;
+	const dvec3 black;
 
 	Cube();
 };
